beacon rate anomaly alert

+++ src/test/java/horse/wtf/nzyme/alerts/CryptoChangeBeaconAlertTest.java
@@ -28,7 +28,7 @@ public class CryptoChangeBeaconAlertTest extends AlertTestHelper {
         assertEquals(a.getMessage(), "SSID [wtf] was advertised with unexpected security settings [WPA2-EAM-PSK-CCMP].");
         assertEquals(a.getType(), Alert.Type.CRYPTO_CHANGE_BEACON);
         assertEquals(a.getSubsystem(), Subsystem.DOT_11);
-        assertEquals(a.getFrameCount(), 1);
+        assertEquals(a.getFrameCount(), (Long) 1L);
         assertFalse(a.getLastSeen().isAfterNow());
         assertTrue(a.getLastSeen().isBeforeNow());
         assertFalse(a.getFirstSeen().isAfterNow());
