+++ src/test/java/horse/wtf/nzyme/alerts/UnexpectedBSSIDBeaconAlertTest.java
@@ -26,7 +26,7 @@ public class UnexpectedBSSIDBeaconAlertTest extends AlertTestHelper {
         assertEquals(a.getMessage(), "SSID [wtf] was advertised with beacon frame by unexpected BSSID [00:c0:ca:95:68:3b].");
         assertEquals(a.getType(), Alert.Type.UNEXPECTED_BSSID_BEACON);
         assertEquals(a.getSubsystem(), Subsystem.DOT_11);
-        assertEquals(a.getFrameCount(), 1);
+        assertEquals(a.getFrameCount(), (Long) 1L);
         assertFalse(a.getLastSeen().isAfterNow());
         assertTrue(a.getLastSeen().isBeforeNow());
         assertFalse(a.getFirstSeen().isAfterNow());
