+++ src/test/java/horse/wtf/nzyme/alerts/UnexpectedFingerprintBeaconAlertTest.java
@@ -31,7 +31,7 @@ public class UnexpectedFingerprintBeaconAlertTest extends AlertTest {
         assertEquals(a.getMessage(), "SSID [wtf] was advertised by a device with unexpected fingerprint [ec398735dc99267d453908d81bfe06ce04cfa2573d0b9edf1d940f0dbf850a9c]");
         assertEquals(a.getType(), Alert.Type.UNEXPECTED_FINGERPRINT_BEACON);
         assertEquals(a.getSubsystem(), Subsystem.DOT_11);
-        assertEquals(a.getFrameCount(), 1);
+        assertEquals(a.getFrameCount(), (Long) 1L);
         assertFalse(a.getLastSeen().isAfterNow());
         assertTrue(a.getLastSeen().isBeforeNow());
         assertFalse(a.getFirstSeen().isAfterNow());
